Consolidate to one logger
